feat(models): improve domain models and services with detailed documentation and new structures
- Added comprehensive documentation to various services including `LangchainCoordinator`, `ChatService`, `IngestorService`, and `QueryFilterService`.
- Introduced new `PipelineService` to orchestrate interactions between services.
- Created new models for `Language`, `Repository`, and `URLs` to improve type safety and clarity.
- Removed deprecated `Pipeline` interface and refactored related components.
- Updated application configuration in `application.conf` for clarity on model options.
