+++ backend/src/test/resources/application.conf
@@ -2,7 +2,7 @@
 gitinsp {
   models {
     provider = "ollama"  # Available options: "Ollama", "Gemini" and "Claude"
-    default-model = "qwen2.5-coder-32b"  # "llama3.3" # "deepseek-r1-32b"
+    default-model = "qwen2.5-coder-32b" # "qwen3-32b", "llama3.3", "deepseek-r1-32b"
   }
   
   ollama {
